pyspur/backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
arxiv==2.1.3
asyncio==3.4.3
atlassian-python-api==3.41.16
attrs==24.3.0
beautifulsoup4==4.12.3
black==24.8.0
blobfile==3.0.0
bokeh==3.6.2
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.8
contourpy==1.3.1
cssselect==1.2.0
dataclasses-json==0.6.7
datasets==3.2.0
Deprecated==1.2.15
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
elevenlabs==1.9.0
fal_client==0.4.1
fastapi==0.115.6
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gkeepapi==0.15.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gpsoauth==1.1.1
greenlet==3.1.1
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
jieba3k==0.35.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.55.12
llama-cloud==0.1.7
llama-hub==0.0.45
llama-index==0.11.12
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.12
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.3.1
llama-index-legacy==0.9.48.post4
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-github==0.2.0
llama-index-readers-google==0.4.1
llama-index-readers-json==0.2.0
llama-index-readers-llama-parse==0.3.0
llama-index-readers-notion==0.2.0
llama-index-readers-reddit==0.2.0
llama-index-readers-twitter==0.2.0
llama-index-readers-web==0.2.2
llama-index-readers-wikipedia==0.2.0
llama-parse==0.5.18
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
marshmallow==3.23.2
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.4.4
openai==1.55.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
playwright==1.49.1
praw==7.8.1
prawcore==2.4.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
psycopg2-binary==2.9.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
PyDrive==1.3.1
pyee==12.0.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.2.0
pypdf==4.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==1.3.1
retrying==1.3.4
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
selenium==4.27.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.0.27
SQLAlchemy==2.0.36
starlette==0.41.3
striprtf==0.0.26
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.3
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
tweepy==4.14.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
update-checker==0.18.0
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
wsproto==1.2.0
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.18.3
zipp==3.21.0